Modified the project a bit. Created a new folder (ModelAndCustomViews), moved anything that are reusable in the app into their separate files (CustomButton, TutorialModel, and RoundedButton), and moved them in the new folder created.
