+++ TechEase/DetailTutorial.swift
@@ -1,0 +1,20 @@
+//
+//  DetailTutorial.swift
+//  TechEase
+//
+//  Created by Natalman Nahm on 5/1/21.
+//
+
+import SwiftUI
+
+struct DetailTutorial: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct DetailTutorial_Previews: PreviewProvider {
+    static var previews: some View {
+        DetailTutorial()
+    }
+}
