+++ TechEase/ModelsAndCustomViews/TutorialModel.swift
@@ -1,0 +1,14 @@
+//
+//  TutorialModel.swift
+//  TechEase
+//
+//  Created by Natalman Nahm on 5/1/21.
+//
+
+import Foundation
+
+struct Tutorial: Identifiable{
+    let id = UUID()
+    let TutorialName: String
+    let Icon: String
+}
