+++ TechEase/TechEaseTutorialList.swift
@@ -8,12 +8,6 @@
 
 import SwiftUI
 
-struct Tutorial: Identifiable{
-    let id = UUID()
-    let TutorialName: String
-    let Icon: String
-}
-
 struct viewTutorial: View {
     let tutorialDisplay: Tutorial
     
@@ -97,20 +91,6 @@ struct TechEaseTutorialList: View {
             }
             
         }
-    }
-}
-
-struct RoundedButton: ButtonStyle {
-    func makeBody(configuration: Self.Configuration) -> some View {
-        return configuration
-            /*
-            Arica: I modified this section to make the buttons look like our high-fidelity prototype. That way, whenever RoundedButton is called, this styling will be used, keeping the design consistent.
-            */
-            .label
-            .padding()
-            .frame(minWidth: /*@START_MENU_TOKEN@*/0/*@END_MENU_TOKEN@*/, maxWidth: /*@START_MENU_TOKEN@*/.infinity/*@END_MENU_TOKEN@*/)
-            .background(RoundedRectangle(cornerRadius: 10).stroke(Color("DarkBlue"), lineWidth: 2).background(Color("LightBlue").cornerRadius(10)))
-            .foregroundColor(.black)
     }
 }
